Add EquippedItem struct
I noticed that the weapon and wearable equipped structs were essentially
identical and while that wasn't bad for just the two structs once I got
into the prototypes and generators it was getting to be quite a bit of
duplicated code. The main issue is now it needs to use a Box for the
generator inside of the generic struct.