Massive equipped item changes and generator
I have swapped to using an enum for equipped location and an actual
value of None instead of it being an Option. This feels a bit like a
side-step and I might remove it eventually, but for right now this means
there's another time to get no equipped location instead of just the
roll.