Remove identifier from character
I've decided to tie the identifiers to non player and player, as I'm
going to be adding an ID to this later. This means characters won't
have identifiers, and will be more about the basic stats of the
character, and the identifier is how the actual character (non player or
player) are identified.