Rework descriptors
This completely reworks how descriptors are done. Now they can be
retrieved by "tags", which are things like "blade", "blunt", "armour",
or materials like "metal" and "wood".

I tried a bunch of ways of going about this, and it was always a little
gross, but this seems maybe the best? I might eventually add to my
prototypes to allow them to filter the possible materials and
descriptors, but this is fine for now.