Create NPC position generator
This generates NPCs for the room. It creates a number of groups randomly
rolled based on the room type, and then picks a group size based on the
species that it starts with. It has a chance of switching species, but
not a high one.

Then it picks a descriptor for the group based on the size and any
fixtures in the room that are applicable.