Add ItemView
This adds my first "view" of an item, and builds it out. Views will be
how I show my components outwards to consumers, probably via commands,
unless they want to use the components directly. But I think the view
will be able to help steer things towards "this is the information the
consumer knows right now".