Move game and add handle to game impl
Game now sits at the root, because it's a very important struct. And it
can now handle actions directly, returning the events that happened
because of the action. It internally calls my handle method with its
state and player, and then it updates its own internal state afterwards.

I'm not in love with game state not necessarily knowing the current
player, but I don't necessarily want to add the player id to it... yet.

Maybe I'll add that in the future just to make the game state a bit more
standalone, but right now the game state assumes that the current player
is kept track of somewhere else.