Remove use of LocationDescriptor
This concept was getting very complicated to keep going as I was
equipping new items, and deciding what to do with it was getting to be
not worth the effort of keeping it going.

Also removing the `Equipped` tag in favour of just a single boolean
flag on the character_item, which is essentially how this was being
used anyways.

Overall I think I might be able to add back the descriptor to the view,
and only keep it there.