Swap equipped location to single
Previously I was letting this be multiple places. I don't think that
really makes sense if this is meant to be where it is currently on the
character. The item (for now) can't both be on their head and feet,
even if the item _can_ be equipped in both places.