Move actions to the game
The game has more context about everything going on, so I think it makes
sense to have it be the thing that decides what the actions that can
happen next are, much moreso than the room.

I'm also starting to add my state around what a player knows about
characters and I hope to expand on this in the near future.