Rework look actions
I am moving to a two action approach for looking at NPCs. By default
when you look at NPCs it will return some basic information. Then the
player will be able to "inspect" an NPC, which lets them try to
determine more information. Right now there are just chances that
the discovery works. Eventually this will change to be able to be
adjusted with spells.