Add player statistics
I want to keep track of player characters and their statistics for this
game state. Because I know that there might be multiple player
characters I have to keep track of them by player character id.

Maybe I'll swap this over for other parts of game state too, but for
now I'm leaving it at this.