Completely rework fixture items
How items were contained in fixtures was way too complicated. Now an
item can either be inside the fixture, inside the hidden compartment,
or neither (so just sitting on top).

And now there's less knowledge necessary for users, they either know
if there's a hidden compartment or not, and they can open the fixture
if it's closed (upcoming).