Create an intermediate state
It was getting to be too hard to manage the events and trying to
figure out an intermediate state while trying to figure out what else
is going on. Instead I just create a new intermediate state/player from
the initial events and then calculate the global effects from that.