Remove items being hidden in inventory
This was something that I had plans for in the very beginning, but it
never really became anything. One day I might revisit but I don't want
code that isn't really used in any real capacity sticking around. I'm
about to add some more options for inventory/npc generation, and as
part of this I want to remove any of the options that would make that
confusing.